src/requirements.txt
discord.py==1.7.3
Pillow==9.1.1
pyboy==1.4.7
python-dotenv==0.20.0
